move constants onto a separate file
